+++ src/annotations/items/AnnotationMarkerRect.h
@@ -32,6 +32,9 @@ public:
 	AnnotationMarkerRect(const AnnotationMarkerRect &other) = default;
 	~AnnotationMarkerRect() override = default;
 	Tools toolType() const override;
+
+protected:
+	void paint(QPainter *painter, const QStyleOptionGraphicsItem *option, QWidget *widget) override;
 };
 
 } // namespace kImageAnnotator